Support for ensuring that UL agents are killable before doing Node kills.
Add this before StartSociety:
do_action "WatchAgentPersists"
-- Note that this takes care of setting the required -D argument

Then, some place before you will be killing nodes, do:
wait_for "ReadyToKillNodes", <node 1>, <node 2>, <etc>

This will block waiting until any agent on any of the listed Nodes
which has an SDClientPlugin has sent the CougaarEvent indicating that
it has persisted after FindProviders.

Alternately, supply no Node names to the wait_for state to wait for
all Nodes in the society to be ready to be killed.

Also note that logic like that in the State could readily be added to
the KillNodes action to combine those 2 steps in one.