Bug 13493 and 13494: 1) Launch a new thread for each Node for advancing time, to try to parallelize the Node accesses 2) Pull out some of the constants used for setting sync time in AdvanceTime 3) Shorten some of the times in AdvanceTime: - Sync time is now 10 seconds sooner: 20+0.5*Nodes becomes 10+0.5*Nodes - When the society is going to go non-Quiescent after advance, it does so quickly. So instead of waiting 20 seconds, wait only 5 seconds.
This saves 25 seconds off each Advance when there is nothing to do.

4) Also use multiple threads for hitting all of the /completion
servlets when saving Agent completion. We use batches of 50 threads to
limit the total threads created. Again, this parallelizes the
SaveSocietyCompletion action.

5) Shorten the time that the ACME quiescence monitor sleeps between
checking for society completion from 10 seconds to 5. It still waits
another 10 to avoid toggling, but this should help it notice
quiescence a little sooner.