Bug 3651: As a work-around for bug 3644, we enabled the new -D argument in GLSExpanderPlugin. The result is that agents persist after Stage-1 GLS has been propogated, _instead_ of after FindProviders. Therefore, GLS-1 will not get rescinded if you kill any time after this persist. Note that no persists occur in later stages, as they are not necessary (rescinding those GLS tasks has no known ill effects on the society).
With this commit, the WatchAgentPersists and
NodesPersistedFindProviders ACME actions will invisibly change to use the
new behavior instead of the old. All scripts will use the new
functionality, without any changes.

To use the old behavior (persist earlier at risk of broken runs when
Nodes are killed), supply the argument "false" to WatchAgentPersists
by editing BaselineTemplate.rb (or changing the default as specified
in this file).